% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesisdowndss.R
\docType{package}
\name{thesisdowndss}
\alias{thesisdowndss}
\alias{thesisdowndss-package}
\title{thesisdowndss: A package for creating Duke Statistical Science undergraduate theses using R Markdown}
\description{
thesisdowndss: A package for creating Duke Statistical Science undergraduate theses using R Markdown
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

